refactor(js): sync goroutines on Babel instance, not Compiler
This is a substantial refactor to essentially allow testing different
Compiler instances that use the same Babel runtime. This was required
in order to test both ES5 and ES6 compilation concurrently, while
loading Babel conditionally only if Compiler.Transform is called.

It also abstracts the Babel API in a nicer way, instead of it being part
of Compiler itself, which I would argue are separate things.