refactor(js): make Compiler invoke a preprocessor, pre-compile core.js
This attempts to further separate the concept of compiler.Compiler and Babel being
the same thing.

A compiler.Compiler does two things:

1. Parses source code into a goja.Program, which in the Goja domain
means *compiling*.
2. Optionally runs the code through preprocessors and pre-compiles shims
such as core.js to be used in the runtime VU Goja VM.

I think introducing the concept of a preprocessor makes sense, as that's
what Babel is (besides being a compiler, which some may disagree with).
This design is more cohesive (all shims are now in js/compiler/lib/, the
js/lib/ package is gone) and gives us a place to expand by further
adding shims of our own.